chore: Fix support for Node.js 18
The initial support for Node.js 18 done in
https://github.com/achrinza/event-pubsub/commit/9e2052d4c36ed735d660d74761659e5b518a014d
was not enough because trying to use this package in Node.js 18 would still fail on the required
version of @achrinza/strong-type dependency, which didn't support node 18.

Note that this solves node 18 support only for production use of this package.
For development in node 18, it also requires to bump the version in
https://github.com/node-ipc/vanilla-test/blob/8e2cfbce5c9a3f98faeb8ae1b7508d4131064cee/package.json#L31
and release a new patch of that package.

Signed-off-by: Vít Stanislav <vits@kryptoslogic.com>